backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=23.5
psycopg2-binary>=2.9
Pillow>=10.0
django-cors-headers>=4.3
pytest>=7.4
pytest-django>=4.7
requests>=2.31
